requirements.txt
pyneuphonic
pyaudio
transformers
torch
python-dotenv